// FamilyClub.React/famiy.club.react/lib/api/generated/models/BookSizeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookSizeDto
 */
export interface BookSizeDto {
    /**
     * 
     * @type {number}
     * @memberof BookSizeDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BookSizeDto
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BookSizeDto
     */
    code?: string | null;
}

/**
 * Check if a given object implements the BookSizeDto interface.
 */
export function instanceOfBookSizeDto(value: object): value is BookSizeDto {
    return true;
}

export function BookSizeDtoFromJSON(json: any): BookSizeDto {
    return BookSizeDtoFromJSONTyped(json, false);
}

export function BookSizeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookSizeDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'code': json['code'] == null ? undefined : json['code'],
    };
}

export function BookSizeDtoToJSON(json: any): BookSizeDto {
    return BookSizeDtoToJSONTyped(json, false);
}

export function BookSizeDtoToJSONTyped(value?: BookSizeDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'code': value['code'],
    };
}